requirements.txt
aiokafka==0.12.0
async-timeout==5.0.1
aws-msk-iam-sasl-signer-python==1.0.2
boto3==1.38.0
botocore==1.38.0
click==8.1.8
confluent-kafka==2.10.0
jmespath==1.0.1
kafka-python==2.1.5
packaging==25.0
python-dateutil==2.9.0.post0
s3transfer==0.12.0
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
